requirements.txt
django
gunicorn
psycopg2
django-crispy-forms
crispy-bootstrap4
django-cors-headers
apscheduler~=3.10.4
google-api-python-client
pytz~=2023.3.post1
Pillow
django-phonenumber-field
phonenumbers